requirements.txt
flask==3.1.0
supabase==2.15.0
stripe==11.4.1
python-dotenv==1.1.0